* Fix a race condition where if the master requested data before the tx buffer was populated, the slave would abort the send and begin waiting for another i2c start condition.
This caused an apparent bus hang when operated against an Arduino using the Wire library because the Wire library performs a blocking read when a slave sends ACK in response to a request to read data. Ideally, the Wire library should time out and program execution should continue, but this is a bug for another project.

The fix causes the slave to return from the USI overflow interrupt handler keeping SCL held low when the master wants to read and the Tx buffer is empty. Processing of the preceeding command will continue until the Tx buffer begins to be populated in which case the main loop will trigger the bytes to be sent.

I have tested the fix on attiny44 and attiny85 hardware at up to and slightly above 400kHz bus speed with the Wire library and have not encountered a single case of the arduino blocking. I have also tested a modified version of the Wire library that allows a repeat start sequence to be sent by the master and the slaves handle that as well.